Add milkdrop to admin player preview.
