+++ src/frontend.js
@@ -1,6 +1,10 @@
 import Webamp from 'webamp';
-import butterchurnPresets from 'butterchurn-presets';
 import domReady from '@wordpress/dom-ready';
+
+/**
+ * Internal dependencies
+ */
+import milkdropOptions from './milkdrop';
 
 domReady( () => {
 	const container = document.querySelector( '.wp-block-tenup-winamp-block' );
@@ -43,28 +47,6 @@ domReady( () => {
 		}
 	}
 
-	options.__butterchurnOptions = {
-		butterchurnOpen: true,
-		importButterchurn: () => import( 'butterchurn' ),
-		getPresets: () => {
-			const presets = butterchurnPresets.getPresets();
-
-			return Object.keys( presets ).map( ( name ) => {
-				return {
-					name,
-					butterchurnPresetObject: presets[ name ],
-				};
-			} );
-		},
-	};
-
-	options.__initialWindowLayout = {
-		main: { position: { x: 0, y: 0 } },
-		equalizer: { position: { x: 0, y: 116 } },
-		playlist: { position: { x: 0, y: 232 }, size: [ 0, 1 ] },
-		milkdrop: { position: { x: 275, y: 0 }, size: [ 7, 9 ] },
-	};
-
 	// Render the player
-	new Webamp( options ).renderWhenReady( container );
+	new Webamp( { ...options, ...milkdropOptions } ).renderWhenReady( container );
 } );
